+++ src/components/LoginComponents/WrapperRight/types.ts
@@ -1,0 +1,9 @@
+export interface ContainerProps {
+	width?: string;
+	height?: string;
+	backgroundColor?: string;
+	display?: string;
+	justifyContent?: string;
+	alignItems?: string;
+	borderRadius?: string;
+}
