+++ src/components/ui-components/Button/index.tsx
@@ -1,0 +1,9 @@
+import { FC, ButtonHTMLAttributes } from 'react'
+
+import { IButtonProps } from './types'
+import * as SC from './styles'
+
+export const Button: React.FC<IButtonProps> = ({ children, ...props }) => (
+	<SC.ButtonWrapper {...props}>{children}</SC.ButtonWrapper>
+)
+
