+++ src/components/ui-components/Input/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+import { InputProps } from './types'
+
+import * as SC from './styles'
+export const Input = (props: InputProps) => <SC.InputWrapper {...props} />
