+++ src/components/ui-components/Line/index.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react'
+import { ILine } from './types'
+import * as SC from './styles'
+
+export const GreyLine: FC<ILine> = ({ ...props }) => <SC.Line {...props} />
