+++ src/components/ui-components/Span/index.tsx
@@ -1,12 +1,12 @@
 import { FC } from 'react'
 import { ISpan } from './types'
-import { SpanWrapper } from './styles'
+import * as SC from './styles'
 
 const Span: FC<ISpan> = ({ text, children, ...props }) => (
-	<SpanWrapper {...props}>
+	<SC.SpanWrapper {...props}>
 		{text}
 		{children}
-	</SpanWrapper>
+	</SC.SpanWrapper>
 )
 
 export default Span
