+++ src/components/ui-components/Text/TitleH1/index.tsx
@@ -1,11 +1,11 @@
 import { FC } from 'react'
 import { ITitleH1 } from './types'
-import { TitleWrapper } from './styles'
+import * as SC from './styles'
 
 const TitleH1: FC<ITitleH1> = ({ text, ...props }) => (
-	<TitleWrapper {...props}>
+	<SC.TitleWrapper {...props}>
 		{text}
-	</TitleWrapper>
+	</SC.TitleWrapper>
 )
 
 export default TitleH1
