+++ src/components/ui-components/Text/TitleInSection/styles.ts
@@ -1,17 +1,16 @@
 import styled from 'styled-components'
 import { ITitleInSection } from './types'
 
-const TitleSectionWrapper = styled.div<ITitleInSection>`
+export const TitleSectionWrapper = styled.div<ITitleInSection>`
   margin-bottom: ${props => props.marginBottom};
   margin-top: ${props => props.marginTop};
   width: ${props => props.width};
 `
 
-const TitleText = styled.text<ITitleInSection>`
+export const TitleText = styled.text<ITitleInSection>`
   font-size: ${props => props.fontSize};
   font-bold: ${props => props.fontBold};
   color: ${props => props.color};
   font-weight: ${props => props.fontWeight};
   font-family: ${props => props.fontFamily};
 `
-export const defaultComponentStyles = { TitleSectionWrapper, TitleText }
