+++ src/pages/login/styles.ts
@@ -9,7 +9,7 @@ export const Container = styled.div<ContainerProps>`
 `
 
 export const Wrapper = styled.div<ContainerProps>`
-  min-width: 94%;
+  width: 95%;
   padding: 20px;
   margin: 20px 40px 20px 40px;
   background-color: ${props => props.theme.colors.blue_light};
@@ -18,35 +18,39 @@ export const Wrapper = styled.div<ContainerProps>`
   align-items: center;
   border-radius: 40px;
 `
-export const WrapperLeft = styled.div`
-  width: 600px;
-  height: 840px;
-  background-color: ${props => props.theme.colors.white};
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 40px;
-  flex-direction: column;
-`
+// export const WrapperLeft = styled.div`
+//   width: 600px;
+//   height: 840px;
+//   background-color: ${props => props.theme.colors.white};
+//   display: flex;
+//   justify-content: center;
+//   align-items: center;
+//   border-radius: 40px;
+//   flex-direction: column;
+//   margin-left: 17px;
+//   box-sizing: border-box;
+// `
 
-export const WrapperRight = styled.div`
-  height: 235px;
-  background-color: ${props => props.theme.colors.blue_light};
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  margin-left: 96px;
-`
+// export const WrapperRight = styled.div`
+//   height: 235px;
+//   background-color: ${props => props.theme.colors.blue_light};
+//   display: flex;
+//   justify-content: center;
+//   align-items: center;
+//   flex-direction: column;
+//   box-sizing: border-box;
+// `
 
-export const Image = styled.img`
-  width: 92px;
-  height: 92px;
-  margin-bottom: 24px;
-`
+// export const Image = styled.img`
+//   width: 92px;
+//   height: 92px;
+//   margin-bottom: 24px;
+// `
 
-export const ButtonsWrapper = styled.div`
-  display: flex;
-  gap: 8px;
-  margin-top: 16px;
-`
+// export const ButtonsWrapper = styled.div`
+//   display: flex;
+//   gap: 8px;
+//   margin-top: 16px;
+//   margin-left: 40px;
+//   margin-right: 40px;
+// `
