+++ src/pages/registration/index.tsx
@@ -2,22 +2,22 @@ import React from 'react'
 import { ButtonsWrapper, Container, Image, Line, Wrapper, WrapperLeft, WrapperRight } from './styles'
 import Logo from '../../assets/logo.svg'
 import ImageLoginPage from '../../assets/svg-pic.png'
-import TitleInSection from '../../ui-components/Text/TitleInSection'
+import TitleInSection from '../../components/ui-components/Text/TitleInSection'
 import Google from '../../assets/GoogleIcon.svg'
 import Github from '../../assets/GithubIcon.svg'
 import LinkedIN from '../../assets/LinkedINicon.svg'
 import Letter from '../../assets/letterInput.svg'
 import LockIcon from '../../assets/lockIcon.svg'
 import PasswordEye from '../../assets/EyeOn.png'
-import { Button } from '../../ui-components/Button'
-import { Block } from '../../ui-components/Block'
-import { Input } from '../../ui-components/Input'
-import Label from '../../ui-components/Label'
-import Span from '../../ui-components/Span'
-import TitleH1 from '../../ui-components/Text/TitleH1'
+import { Button } from '../../components/ui-components/Button'
+import { Block } from '../../components/ui-components/Block'
+import { Input } from '../../components/ui-components/Input'
+import Label from '../../components/ui-components/Label'
+import Span from '../../components/ui-components/Span'
+import TitleH1 from '../../components/ui-components/Text/TitleH1'
 import { theme } from '../../theme/theme'
-import { GreyLine } from '../../ui-components/Line'
-import Link from '../../ui-components/Route'
+import { GreyLine } from '../../components/ui-components/Line'
+import Link from '../../components/ui-components/Route'
 
 export const Registration = () => {
