--- src/ui-components/Button/index.tsx
@@ -1,10 +1,0 @@
-import { FC, ButtonHTMLAttributes } from 'react'
-import { ButtonWrapper } from './styles'
-import { IButtonProps } from './types'
-
-// interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {}
-
-export const Button: React.FC<IButtonProps> = ({ children, ...props }) => (
-	<ButtonWrapper {...props}>{children}</ButtonWrapper>
-)
-
