--- src/ui-components/Input/index.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-import { InputProps } from './types'
-import { InputWrapper } from './styles'
-
-export const Input = (props: InputProps) => <InputWrapper {...props} />
