--- src/ui-components/Line/index.tsx
@@ -1,5 +1,0 @@
-import { FC } from 'react'
-import { ILine } from './types'
-import { Line } from './styles'
-
-export const GreyLine: FC<ILine> = ({ ...props }) => <Line {...props} />
